+++ neoforge/src/main/java/com/iamkaf/bonded/neoforge/datagen/DataGenerators.java
@@ -58,5 +58,7 @@ public class DataGenerators {
                 event.includeClient(),
                 new ModBlockStateProvider(packOutput, existingFileHelper)
         );
+
+        generator.addProvider(event.includeClient(), new ModSoundProvider(packOutput, existingFileHelper));
     }
 }